+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Bug fixes
   * Fix a bug where a dependency wasn't detected as overridden if was being overridden in another branch of the dependency tree
+  * Fix a bug where a dependency converged in mix did not consider all its requirements
 
 ## v0.7.1 (2015-02-15)